# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.MatricularRequest:
    properties:
      trilha_id:
        type: integer
      usuario_id:
        type: integer
    required:
    - trilha_id
    - usuario_id
    type: object
  model.CreateTrilhaRequest:
    properties:
      carga_horaria:
        type: integer
      descricao:
        type: string
      foco_principal:
        maxLength: 100
        type: string
      nivel:
        enum:
        - INICIANTE
        - INTERMEDIARIO
        - AVANCADO
        type: string
      nome:
        maxLength: 150
        minLength: 5
        type: string
    required:
    - carga_horaria
    - nivel
    - nome
    type: object
  model.CreateUsuarioRequest:
    properties:
      area_atuacao:
        maxLength: 100
        type: string
      email:
        type: string
      nivel_carreira:
        maxLength: 50
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - nome
    type: object
  model.ErrorResponse:
    properties:
      details:
        type: string
      message:
        type: string
    type: object
  model.Matricula:
    properties:
      data_inscricao:
        type: string
      id:
        type: integer
      status:
        description: ATIVA, CONCLUIDA, CANCELADA
        type: string
      trilha_id:
        type: integer
      usuario_id:
        type: integer
    type: object
  model.TrilhaResponse:
    properties:
      carga_horaria:
        type: integer
      descricao:
        type: string
      foco_principal:
        type: string
      id:
        type: integer
      nivel:
        type: string
      nome:
        type: string
    type: object
  model.UpdateTrilhaRequest:
    properties:
      carga_horaria:
        type: integer
      descricao:
        type: string
      foco_principal:
        maxLength: 100
        type: string
      nivel:
        enum:
        - INICIANTE
        - INTERMEDIARIO
        - AVANCADO
        type: string
      nome:
        maxLength: 150
        minLength: 5
        type: string
    type: object
  model.UpdateUsuarioRequest:
    properties:
      area_atuacao:
        maxLength: 100
        type: string
      nivel_carreira:
        maxLength: 50
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
    type: object
  model.UsuarioResponse:
    properties:
      area_atuacao:
        type: string
      data_cadastro:
        type: string
      email:
        type: string
      id:
        type: integer
      nivel_carreira:
        type: string
      nome:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@manus.im
    name: Manus AI
    url: https://help.manus.im
  description: API RESTful para uma plataforma de Upskilling/Reskilling voltada ao
    futuro do trabalho (2030+).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Plataforma de Upskilling/Reskilling API
  version: "1.0"
paths:
  /matriculas:
    post:
      consumes:
      - application/json
      description: Realiza a inscrição de um usuário em uma trilha de aprendizagem.
      parameters:
      - description: Dados da Matrícula
        in: body
        name: matricula
        required: true
        schema:
          $ref: '#/definitions/controller.MatricularRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Matricula'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Matricular usuário em uma trilha
      tags:
      - Matriculas
  /trilhas:
    get:
      description: Retorna uma lista de todas as trilhas cadastradas.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TrilhaResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Lista todas as trilhas de aprendizagem
      tags:
      - Trilhas
    post:
      consumes:
      - application/json
      description: Cria uma nova trilha de upskilling/reskilling.
      parameters:
      - description: Dados da Trilha
        in: body
        name: trilha
        required: true
        schema:
          $ref: '#/definitions/model.CreateTrilhaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TrilhaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova trilha de aprendizagem
      tags:
      - Trilhas
  /trilhas/{id}:
    delete:
      description: Remove uma trilha de aprendizagem.
      parameters:
      - description: ID da Trilha
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Deleta uma trilha
      tags:
      - Trilhas
    get:
      description: Retorna os detalhes de uma trilha específica.
      parameters:
      - description: ID da Trilha
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrilhaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca uma trilha por ID
      tags:
      - Trilhas
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma trilha existente.
      parameters:
      - description: ID da Trilha
        in: path
        name: id
        required: true
        type: integer
      - description: Dados da Trilha para atualização
        in: body
        name: trilha
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTrilhaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrilhaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma trilha
      tags:
      - Trilhas
  /usuarios:
    get:
      description: Retorna uma lista de todos os usuários cadastrados.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UsuarioResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Lista todos os usuários
      tags:
      - Usuarios
    post:
      consumes:
      - application/json
      description: Cria um novo usuário na plataforma de upskilling/reskilling.
      parameters:
      - description: Dados do Usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.CreateUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo usuário
      tags:
      - Usuarios
  /usuarios/{id}:
    delete:
      description: Remove um usuário da plataforma.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Deleta um usuário
      tags:
      - Usuarios
    get:
      description: Retorna os detalhes de um usuário específico.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UsuarioResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca um usuário por ID
      tags:
      - Usuarios
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário existente.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do Usuário para atualização
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UsuarioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um usuário
      tags:
      - Usuarios
  /usuarios/{id}/matriculas:
    get:
      description: Retorna todas as matrículas de um usuário específico.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Matricula'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Lista matrículas de um usuário
      tags:
      - Matriculas
swagger: "2.0"